requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.8.0

# Detection
ultralytics>=8.0.0

# Tracking
filterpy>=1.4.5
lap>=0.4.0

# Re-ID
torchreid>=0.2.5

# Evaluation
motmetrics>=1.4.0

# Configuration
pyyaml>=6.0
easydict>=1.10

# Visualization & GUI
matplotlib>=3.7.0
PySide6>=6.5.0

# Utilities
tqdm>=4.65.0
pandas>=2.0.0
loguru>=0.7.0
tabulate>=0.9.0

# Weights download (Re-ID model zoo lives on Google Drive)
gdown>=4.7.0
